movies_admin/requirements.txt
Django==4.2.11

Faker==9.2.0
psycopg2-binary==2.9.1
python-dotenv==1.0.1
django-split-settings==1.3.2
django-debug-toolbar==4.4.6